[NO-TICKET] refactored controllers/api to be dependent on api spec
