--- backend/src/main/java/org/minttwo/api/account/GetAccountResponseDto.java
@@ -1,11 +1,0 @@
-package org.minttwo.api.account;
-
-import lombok.Builder;
-import lombok.Data;
-
-
-@Builder
-@Data
-public class GetAccountResponseDto {
-    private AccountDto account;
-}
