--- backend/src/main/java/org/minttwo/api/account/GetAccountTransactionResponseDto.java
@@ -1,10 +1,0 @@
-package org.minttwo.api.account;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class GetAccountTransactionResponseDto {
-    private AccountTransactionDto accountTransaction;
-}
