--- backend/src/main/java/org/minttwo/api/account/ListAccountTransactionsResponseDto.java
@@ -1,12 +1,0 @@
-package org.minttwo.api.account;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-
-@Builder
-@Data
-public class ListAccountTransactionsResponseDto {
-    private List<AccountTransactionDto> accountTransactionDtoList;
-}
