--- backend/src/main/java/org/minttwo/api/account/ListAccountsResponseDto.java
@@ -1,12 +1,0 @@
-package org.minttwo.api.account;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-
-@Builder
-@Data
-public class ListAccountsResponseDto {
-    private List<AccountDto> accounts;
-}
