--- backend/src/main/java/org/minttwo/api/account/adapters/AccountAdapter.java
@@ -1,16 +1,0 @@
-package org.minttwo.api.account.adapters;
-
-import lombok.NonNull;
-import org.minttwo.api.account.AccountDto;
-import org.minttwo.models.Account;
-
-public class AccountAdapter {
-    @NonNull
-    public AccountDto adapt(@NonNull Account account) {
-        return AccountDto.builder()
-                .id(account.getId())
-                .userId(account.getUserId())
-                .balance(account.getBalance())
-                .build();
-    }
-}
