+++ backend/src/main/java/org/minttwo/api/exception/AccessDeniedException.java
@@ -1,0 +1,10 @@
+package org.minttwo.api.exception;
+
+import jakarta.annotation.Nullable;
+import lombok.NonNull;
+
+public class AccessDeniedException extends RuntimeException {
+    public AccessDeniedException(@NonNull String message, @Nullable Exception e) {
+        super(message, e);
+    }
+}
