+++ backend/src/main/java/org/minttwo/api/exception/NotFoundException.java
@@ -1,0 +1,10 @@
+package org.minttwo.api.exception;
+
+import jakarta.annotation.Nullable;
+import lombok.NonNull;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(@NonNull String message, @Nullable Exception e) {
+        super(message, e);
+    }
+}
