--- backend/src/main/java/org/minttwo/api/user/GetUserResponseDto.java
@@ -1,10 +1,0 @@
-package org.minttwo.api.user;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class GetUserResponseDto {
-    private UserDto user;
-}
