--- backend/src/main/java/org/minttwo/api/user/LoginUserRequestDto.java
@@ -1,11 +1,0 @@
-package org.minttwo.api.user;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class LoginUserRequestDto {
-    private String username;
-    private String password;
-}
