--- backend/src/main/java/org/minttwo/api/user/UserDto.java
@@ -1,12 +1,0 @@
-package org.minttwo.api.user;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class UserDto {
-    private String id;
-    private String email;
-    private String username;
-}
