+++ backend/src/main/java/org/minttwo/data/dataclients/Db.java
@@ -1,0 +1,21 @@
+package org.minttwo.data.dataclients;
+
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+
+public class Db {
+    private final SessionFactory sessionFactory;
+    private Session currentSession;
+
+    public Db(SessionFactory sessionFactory) {
+        this.sessionFactory = sessionFactory;
+    }
+
+    public Session getCurrentSession() {
+        if (currentSession == null) {
+            this.currentSession = sessionFactory.openSession();
+        }
+
+        return currentSession;
+    }
+}
