+++ backend/src/main/java/org/minttwo/data/models/AccountTransactionModel.java
@@ -1,0 +1,39 @@
+package org.minttwo.data.models;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "account_transaction")
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class AccountTransactionModel {
+    @Id
+    @Column(name = "id")
+    private String id;
+
+    @Column(name = "account_id")
+    private String accountId;
+
+    @Column(name = "title")
+    private String title;
+
+    @Column(name = "amount")
+    private Double amount;
+
+    @Column(name = "created_at")
+    private LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+}
