+++ backend/src/main/java/org/minttwo/data/validators/UserValidator.java
@@ -1,0 +1,19 @@
+package org.minttwo.data.validators;
+
+import lombok.NonNull;
+import org.minttwo.api.exception.BadRequestException;
+import org.minttwo.data.models.UserModel;
+
+public class UserValidator {
+    public void validate(@NonNull UserModel userModel) {
+        if (userModel.getEmail() == null || userModel.getEmail().isEmpty()) {
+            throw new BadRequestException("Email is required and cannot be blank", null);
+        }
+        if (userModel.getUsername() == null || userModel.getUsername().isEmpty()) {
+            throw new BadRequestException("Username is required and cannot be blank", null);
+        }
+        if (userModel.getPassword() == null || userModel.getPassword().isEmpty()) {
+            throw new BadRequestException("Password is required and cannot be blank", null);
+        }
+    }
+}
