--- backend/src/main/java/org/minttwo/exception/AccessDeniedException.java
@@ -1,10 +1,0 @@
-package org.minttwo.exception;
-
-import jakarta.annotation.Nullable;
-import lombok.NonNull;
-
-public class AccessDeniedException extends RuntimeException {
-    public AccessDeniedException(@NonNull String message, @Nullable Exception e) {
-        super(message, e);
-    }
-}
