--- backend/src/main/java/org/minttwo/exception/BadRequestException.java
@@ -1,10 +1,0 @@
-package org.minttwo.exception;
-
-import jakarta.annotation.Nullable;
-import lombok.NonNull;
-
-public class BadRequestException extends RuntimeException {
-    public BadRequestException(@NonNull String message, @Nullable Exception e) {
-        super(message, e);
-    }
-}
