--- backend/src/main/java/org/minttwo/exception/NotFoundException.java
@@ -1,10 +1,0 @@
-package org.minttwo.exception;
-
-import jakarta.annotation.Nullable;
-import lombok.NonNull;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(@NonNull String message, @Nullable Exception e) {
-        super(message, e);
-    }
-}
