--- backend/src/main/java/org/minttwo/models/Account.java
@@ -1,36 +1,0 @@
-package org.minttwo.models;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "account")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class Account {
-    @Id
-    @Column(name = "id")
-    private String id;
-
-    @Column(name = "user_id")
-    private String userId;
-
-    @Column(name = "balance")
-    private Double balance;
-
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-}
