+++ backend/src/test/java/org/minttwo/controllers/UserControllerTest.java
@@ -2,15 +2,11 @@ package org.minttwo.controllers;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.minttwo.api.user.GetUserResponseDto;
-import org.minttwo.api.user.LoginUserRequestDto;
-import org.minttwo.api.user.UserDto;
-import org.minttwo.controllers.user.UserController;
-import org.minttwo.dataclients.UserClient;
-import org.minttwo.exception.AccessDeniedException;
-import org.minttwo.exception.BadRequestException;
-import org.minttwo.exception.NotFoundException;
-import org.minttwo.models.User;
+import org.minttwo.data.dataclients.UserClient;
+import org.minttwo.api.exception.AccessDeniedException;
+import org.minttwo.api.exception.BadRequestException;
+import org.minttwo.api.exception.NotFoundException;
+import org.minttwo.data.models.User;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.InjectMocks;
